Move utils functions to separate file
