+++ release/files_tici
@@ -9,6 +9,8 @@ selfdrive/assets/training_wide/*
 
 system/camerad/cameras/camera_qcom2.cc
 system/camerad/cameras/camera_qcom2.h
+system/camerad/cameras/camera_util.cc
+system/camerad/cameras/camera_util.h
 system/camerad/cameras/real_debayer.cl
 
 selfdrive/sensord/rawgps/*
